# Avoid modifying unnecessary nodes for performance