--- a/sig/activerecord/originator/arel_visitor_extension.rbs
+++ b/sig/activerecord/originator/arel_visitor_extension.rbs
@@ -1,6 +1,7 @@
 module ActiveRecord
   module Originator
     module ArelVisitorExtension : _ArelVisitor
+      TARGET_NODE_CLASSESS: Array[Symbol]
       def accept: (untyped object, _Collector collector) -> untyped
 
       private